Homework 13 11 Adding database to the service

+++ src/OzonEdu.MerchendiseService.Domain/Models/Enumeration.cs
@@ -9,9 +9,9 @@ namespace OzonEdu.MerchendiseService.Domain.Models
     {
         public string Name { get; private set; }
 
-        public int Id { get; private set; }
+        public long Id { get; private set; }
 
-        protected Enumeration(int id, string name) => (Id, Name) = (id, name);
+        protected Enumeration(long id, string name) => (Id, Name) = (id, name);
 
         public override string ToString() => Name;
